requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask
Flask-PyMongo
flask-sqlalchemy
flask-migrate
flask-restful